+++ src/SwarmFit/Extensions.cs
@@ -1,0 +1,6 @@
+﻿namespace SwarmFit;
+
+public static class Extensions
+{
+    
+}