+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * SwarmFitter: Improved performance by minimizing allocations (#1) @neon-sunset
 * Renamed classes and arguments to favor `parameters` instead of `variables`
 * SwarmFitter: Use a customizable `IRandomNumberGenerator` with a default that maximizes performance
+* IFunction: Created an interface to help display documented common functions with known typical parameters
 
 # SwarmFit 0.0.2
 _Published to NuGet on 2024-09-14_